+++ tests/numpy/sort_test.py
@@ -36,7 +36,7 @@ class TestSort:
         assert result.failed == 0, f"Doctest failed: {result.failed} failures"
 
     @pytest.mark.parametrize("size", pytest.prob_size)
-    @pytest.mark.parametrize("dtype", [ak.float64, ak.uint64, ak.float64])
+    @pytest.mark.parametrize("dtype", [ak.int64, ak.uint64, ak.float64])
     def test_compare_argsort(self, size, dtype):
         # create np version
         a = np.arange(size, dtype=dtype)
@@ -51,7 +51,7 @@ class TestSort:
         assert np.array_equal(a, b.to_ndarray())
 
     @pytest.mark.parametrize("size", pytest.prob_size)
-    @pytest.mark.parametrize("dtype", [ak.float64, ak.uint64, ak.float64])
+    @pytest.mark.parametrize("dtype", [ak.int64, ak.uint64, ak.float64])
     def test_compare_sort(self, size, dtype):
         # create np version
         a = np.arange(size, dtype=dtype)
